+++ .env
@@ -1,7 +1,8 @@
 PORT=3333
 
-#bancodedados 
+
+#banco de dados 
 DB_HOST=127.0.0.1
-DB_NAME=achadosapp
+DB_NAME=igordb
 DB_USER=root
 DB_PASS=
