+++ .gitignore
@@ -1,8 +1,15 @@
 /node_modules
+
 /build
+
 /dist
 
+yarn.lock
+
+package-lock.json
+
 .env.local
+
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
