+++ nodemon.json
@@ -1,4 +1,5 @@
-{ "execMap":{
-    "js": "sucrase-node"
+{
+    "execMap": {
+        "js": "sucrase-node"
     }
 }
