+++ src/app-modeclass.js
@@ -1,0 +1,26 @@
+import express from "express";
+import cors from "cors";
+import routes from "./routes.js";
+import "./database/index"
+
+
+class App {
+    constructor() {
+        this.server = express();
+        this.middlewares();
+        this.routes();
+    }
+    middlewares() {
+        this.server.use(cors());
+        this.server.use(express.json());
+
+    }
+
+
+
+    routes() {
+        this.server.use(routes);
+    }
+
+}
+export default new App().server
