--- src/app/Controllers/HomeController.js
@@ -1,9 +1,0 @@
-class HomeController{
-
-    async index(req, res){
-        return res.status(200).json({ok: "Hello World"});
-    }
-
-}
-
-export default new HomeController();
