+++ src/app/models/User.js
@@ -1,13 +1,13 @@
 import Sequelize, { Model } from 'sequelize';
 
 class User extends Model {
+
   static init(sequelize) {
     super.init({
       name: Sequelize.STRING,
       mail: Sequelize.STRING,
       age: Sequelize.INTEGER,
-      pass: Sequelize.STRING,      
-    
+      pass: Sequelize.STRING,
     },
       {
         sequelize,
@@ -18,5 +18,8 @@ class User extends Model {
      return this;
 
   }
+ 
+
 }
+
 export default User;
