+++ src/config/database.js
@@ -1,6 +1,7 @@
 "use strict";
 require('dotenv/config');
 
+
 module.exports = {
   dialect: 'mysql',
   host: process.env.DB_HOST,
