+++ src/database/seeds/20200930141200-users.js
@@ -3,10 +3,15 @@
 module.exports = {
   up: async (queryInterface, Sequelize) => {
     return queryInterface.bulkInsert('cad_users', [
+
+
      { name: 'admin', mail: 'admin@mail.com', age:29, pass: '123456', created_at: new Date(), updated_at: new Date()},
+
+
     ], {});
     
   },
+
   down: async (queryInterface, Sequelize) => {
     return queryInterface.bulkDelete('cad_users', null, {});
   }
