+++ src/server.js
@@ -1,6 +1,5 @@
 import app from "./app.js";
 
-// setagem de porta
-
 const PORT = process.env.PORT || 3333;
+
 app.listen(PORT);
